Adding filter for UIKit
Good work, 
As I understand, you should have a filter on the processes running UIKit. This is to reduce the amount of times that UIKit is called unnecessarily. Thus wasting battery and resources.
You can see an example of this on this open source code by litten. 
Also, I would love to work with you on this if you're up to it!